refactor(Alarms): Changed name that contains 'ChartAlarm...' to 'AlarmsGraph...'
