+++ strateger-react/src/components/Alarms/components/AlarmInfoPanel/AlarmsGraphByMonth/hooks/useGenerateChartData.jsx
