+++ strateger-react/src/components/Alarms/components/AlarmInfoPanel/AlarmsGraphByMonth/utils/initChart.jsx
