+++ strateger-react/src/components/Alarms/components/AlarmInfoPanel/AlarmsGraphByTime/utils/configChart.jsx
